Title: 세 실수의 반올림, Time: 137ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240626/세 실수의 반올림/rounding-of-three-actual-numbers.py
@@ -1,5 +1,5 @@
-a = int(input())
-b = int(input())
-c = int(input())
+a = float(input())
+b = float(input())
+c = float(input())
 
 print(f"{a:.3f}\n{b:.3f}\n{c:.3f}")
